reorganize views into multiple files
