+++ .gitignore
@@ -3,4 +3,4 @@ venv/
 __pycache__/
 db.sqlite3
 migrations/
-ten_k_brews_app/.env
+ten_k_brews_app/views/.env
