+++ ten_k_brews_app/views/detail_views.py
@@ -1,12 +1,9 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth import authenticate, login, logout
-from django.contrib.auth.models import User
-from .models import Establishment, Drink, UserData
-from .forms import EstablishmentSearchForm, NewDrinkForm, UserRegistrationForm
+from ..models import Establishment, Drink, UserData
+from ..forms import EstablishmentSearchForm, NewDrinkForm, UserRegistrationForm
+from .. import osm_geolocator
 import environ
-from . import osm_geolocator
 
 search_form = EstablishmentSearchForm   # for search bar used in title_bar.html
 
@@ -14,38 +11,6 @@ search_form = EstablishmentSearchForm   # for search bar used in title_bar.html
 env = environ.Env()
 environ.Env.read_env()
 mapbox_token = env('MAPBOX_TOKEN')
-
-
-def home(request):
-    return render(request, 'home.html', {'search_form': search_form})
-
-
-def browse(request, type_filter):
-    # get establishments to list based on user selection
-    if type_filter == 'all':
-        establishments = Establishment.objects.all().order_by('name')
-    elif type_filter == 'breweries':
-        establishments = Establishment.objects.filter(type=Establishment.BREWERY).order_by('name')
-    elif type_filter == 'wineries':
-        establishments = Establishment.objects.filter(type=Establishment.WINERY).order_by('name')
-    elif type_filter == 'distilleries':
-        establishments = Establishment.objects.filter(type=Establishment.DISTILLERY).order_by('name')
-    elif type_filter == 'cideries':
-        establishments = Establishment.objects.filter(type=Establishment.CIDERY).order_by('name')
-    else:
-        return redirect('home')     # redirect home if type_filter is something unexpected
-
-    return render(request, 'browse_pages/list.html', {'establishments': establishments, 'search_form': search_form})
-
-
-def search(request):
-    search_term = request.GET.get('search_term')
-
-    if search_term:     # check that search term is entered
-        establishments = Establishment.objects.filter(name__icontains=search_term).order_by('name')
-        return render(request, 'browse_pages/list.html', {'establishments': establishments, 'search_form': search_form})
-
-    return redirect('browse', type_filter='all')     # show /browse/all if no search term provided
 
 
 def establishment_detail(request, establishment_pk):
@@ -143,49 +108,3 @@ def new_drink_form(request, establishment_pk):
 
     return render(request, 'form_pages/new_drink.html',
                   {'drink_form': drink_form, 'establishment': establishment, 'search_form': search_form})
-
-
-def user_profile(request, username):
-    user = User.objects.get_by_natural_key(username)
-    user_data = UserData.objects.get(user=user)
-
-    places_visited = user_data.user_establishments.all().order_by('name')
-    drinks_added = Drink.objects.filter(user=user).order_by('name')
-    drinks_drunk = user_data.user_drinks.all().order_by('name')
-
-    return render(request, 'account_pages/user_profile.html',
-                  {'user': user, 'search_form': search_form,
-                   'places_visited': places_visited, 'drinks_drunk': drinks_drunk, 'drinks_added': drinks_added})
-
-
-def register(request):
-    if request.method == 'POST':
-        registration_form = UserRegistrationForm(request.POST)
-        if registration_form.is_valid():
-            user = registration_form.save()
-            user = authenticate(username=request.POST['username'], password=request.POST['password1'])
-
-            # make and save data object for the user
-            user_data = UserData(user=user)
-            user_data.save()
-
-            if user:
-                login(request, user)
-                return redirect('user_profile', username=user.username)
-            else:
-                messages.add_message(request, messages.ERROR, 'Unable to log in new user')
-        else:
-            messages.add_message(request, messages.INFO, 'Please check the data you entered')
-            # include invalid form with error messages added to it. Error messages will be displayed by the template.
-            return render(request, 'account_pages/register.html',
-                          {'form': registration_form, 'search_form': search_form})
-
-    registration_form = UserRegistrationForm()
-    return render(request, 'account_pages/register.html',
-                  {'registration_form': registration_form, 'search_form': search_form})
-
-
-def logout_user(request):
-    username = request.user.username
-    logout(request)
-    return render(request, 'account_pages/logout.html', {'username': username, 'search_form': search_form})
